Fix task 102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,13 +6,13 @@
   */
 int main(void)
 {
-int num1 = 0;
-int num2 = 1;
-int aux, i;
+long num1 = 0;
+long num2 = 1;
+long aux, i;
 for (i = 0; i < 50; i++)
 {
 aux = num1 + num2;
-printf("%d", aux);
+printf("%ld", aux);
 if (i != 49)
 putchar(',');
 num1 = num2;
